+++ Assets/Scripts/PlayerManager.cs
@@ -10,6 +10,9 @@ public class PlayerManager : MonoBehaviour
     [SerializeField]
     private Transform[] slots;
 
+    [SerializeField]
+    private ScoreCounter[] counter;
+
     private Transform[] laneBot = new Transform[5];
     private Transform[] laneMid = new Transform[5];
     private Transform[] laneTop = new Transform[3];
@@ -42,17 +45,24 @@ public class PlayerManager : MonoBehaviour
 
     public void MoveMonstersToField()
     {
+        // Clear Monster from triggering previous game
+        foreach (ScoreCounter _counter in counter)
+        {
+            _counter.ClearList();
+        }
+        
+        // Move Monsters to Players Field
         for (int i = 0; i < myMonsters.Count; i++)
         {
-            myMonsters[i].transform.SetParent(slots[i]);
+            //myMonsters[i].transform.SetParent(slots[i]);
             myMonsters[i].SetActive(true);
-            myMonsters[i].transform.SetAsFirstSibling();
-            myMonsters[i].transform.localPosition = Vector3.zero;
+            //myMonsters[i].transform.SetAsFirstSibling();
+            myMonsters[i].transform.position = slots[i].position;
         }
 
     }
 
-    public void KillAllMonstersFromField()
+    public void RemoveMonstersFromField()
     {
         for (int i = 0; i < myMonsters.Count; i++)
         {
@@ -64,246 +74,21 @@ public class PlayerManager : MonoBehaviour
 
     public void CountScore()
     {
-        updateEachLane();
-        
-        counting();
-
-        PUi.UpdatePlayerUI();
+        CountCounterList();
 
 
     }
 
-    #region Scoring method
-
-    private void counting()
+    private void CountCounterList()
     {
-        elementCounter(laneBot, 0);
-        elementCounter(laneMid, 1);
-        elementCounter(laneTop, 2);
-
-        // Counting Score based on Lane
-        totalScore[0] = TotalOn(laneBot, 0);
-        totalScore[1] = TotalOn(laneMid, 1);
-        totalScore[2] = TotalOn(laneTop, 2);
-
-        
-    }
-
-    int TotalOn(Transform[] _lane, int _laneIndex)
-    {
-        int _result = 0;
-        int _resultElement = 0;
-        int _resultTier = 0;
-        int _resultPair = 0;
-        int _resultAce = 0;
-        int _resultStrongElement = 0;
-
-        for (int i = 0; i < _lane.Length; i++)
+        foreach (ScoreCounter _counter in counter)
         {
-            Element _monster = _lane[i].GetChild(0).GetComponent<Element>();
-
-            
-            // 5 Element pair with Strong type
-            if (_monster.m_Emotion == EmotionType.Strong)
-            {
-                ElementType strongType = _monster.m_ElementType;
-
-                if(checkAllSameElement(_lane, strongType))
-                {
-                    _resultStrongElement = 5;
-                }
-
-
-            }
-
-            // 1 lane same element with no Strong type
-            if (checkAllSameElement(_lane, ElementType.fire))
-            {
-                _resultElement = AllElementPoint;
-            }
-            else if (checkAllSameElement(_lane, ElementType.air))
-            {
-                _resultElement = AllElementPoint;
-            }
-            else if (checkAllSameElement(_lane, ElementType.earth))
-            {
-                _resultElement = AllElementPoint;
-            }
-            else if (checkAllSameElement(_lane, ElementType.water))
-            {
-                _resultElement = AllElementPoint;
-            }
-
-            //All element
-            if (allElement(_laneIndex))
-            {
-                _resultElement = AllElementPoint;
-            }
-
-            // Tier
-            _resultTier += _monster.m_Tier;
-
-
-            
-
-            //_result = _resultTier + _resultElement + _resultPair + _resultAce;
-
-
-            
-            
-        }
-
-        for (int c = 0; c < _lane.Length; c++)
-        {
-            Element _monster = _lane[c].GetChild(0).GetComponent<Element>();
-            // Ace with other Element
-            if (_monster.m_Emotion == EmotionType.Ace)
-            {
-                ElementType AceType = _monster.m_ElementType;
-
-                if (checkAllSameElement(_lane, AceType))
-                {
-                    _resultAce = AcePoint;
-
-                }
-                else
-                {
-                    _resultAce = -AcePoint/2;
-
-                }
-
-
-                _resultPair = 0;
-                _resultElement = 0;
-
-
-            }
-            // No ace no Strong, so using pairing
-            else if (_monster.m_Emotion != EmotionType.Strong)
-            {
-                // 3 Element pair on Bot and Mid lane
-                if (_laneIndex < 2)
-                {
-                    _resultPair = pair3and4(_laneIndex);
-
-                }
-            }
-        }
-
-        _result = _resultElement + _resultPair + _resultAce + _resultStrongElement;
-
-
-        if (_laneIndex == 1)
-            Debug.Log("Tier:" + _resultTier + "  Element:" + _resultElement + "  Pair:" + _resultPair + "  Strong:" + _resultStrongElement + "  Ace:" + _resultAce);
-
-        return _result;
-    }
-
-    bool allElement(int _laneIndex)
-    {
-        if (eFire[_laneIndex] > 0 && eAir[_laneIndex] > 0 && eEarth[_laneIndex] > 0 && eWater[_laneIndex] > 0)
-        {
-            return true;
-        }
-
-        return false;
-    }
-
-    int pair3and4(int _laneIndex)
-    {
-        int _result = 0;
-
-        if (eFire[_laneIndex] == 3)
-        {
-            _result = Pair3Point;
-        }
-        else if (eFire[_laneIndex] == 4)
-        {
-            _result = Pair3Point + 2;
-        }
-
-        if (eAir[_laneIndex] == 3)
-        {
-            _result = Pair3Point;
-        }
-        else if (eAir[_laneIndex] == 4)
-        {
-            _result = Pair3Point + 2;
-        }
-
-        if (eEarth[_laneIndex] == 3)
-        {
-            _result = Pair3Point;
-        }
-        else if (eEarth[_laneIndex] == 4)
-        {
-            _result = Pair3Point + 2;
-        }
-
-        if (eWater[_laneIndex] == 3)
-        {
-            _result = Pair3Point;
-        }
-        else if (eWater[_laneIndex] == 4)
-        {
-            _result = Pair3Point +2;
-        }
-
-        return _result;
-    }
-
-    void elementCounter(Transform[] _lane, int _laneIndex)
-    {
-        eFire[_laneIndex] = 0;
-        eEarth[_laneIndex] = 0;
-        eAir[_laneIndex] = 0;
-        eWater[_laneIndex] = 0;
-
-        for (int i = 0; i < _lane.Length; i++)
-        {
-            Element _monster = _lane[i].GetChild(0).GetComponent<Element>();
-
-            switch (_monster.m_ElementType)
-            {
-                case ElementType.fire:
-                    eFire[_laneIndex]++;
-                    break;
-                case ElementType.water:
-                    eWater[_laneIndex]++;
-                    break;
-                case ElementType.air:
-                    eAir[_laneIndex]++;
-                    break;
-                case ElementType.earth:
-                    eEarth[_laneIndex]++;
-                    break;
-                default:
-                    break;
-            }
+            _counter.Counting();
         }
     }
-
-    bool checkAllSameElement(Transform[] _lane, ElementType _type)
-    {
-        for (int i = 0; i < _lane.Length; i++)
-        {
-            Element _monster = _lane[i].GetChild(0).GetComponent<Element>();
-
-            if(_monster.m_ElementType != _type)
-            {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    #endregion
 
     private void updateEachLane()
     {
-        
-
         int a = 0, b = 0, c = 0;
 
         for (int i = 0; i < slots.Length; i++)
