+++ Assets/Scripts/ScoreCounter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0df10c11975f07d4ab5e32b6504e509a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
